requirements.txt
streamlit==1.28.1
python-pptx==0.6.21
PyPDF2==3.0.1
python-dotenv==1.0.0
pyyaml==6.0
jinja2==3.1.2
msal==1.23.0
pandas==2.0.3
